Change how attention to each am ac is computed - II
Instead of saving and returning three tensors, return one.
At the same time, compute and return att_head_sum to compute forward in the pipeline the percentage attention. Percentage attention was stored before in the two tensors now deleted.
Before staging L_att_to_am_ac to a tensor, insert zeros in the list where there are missing types of amino acids with respect to the complete set of twenty. In this way, later will be possible to sum the attention to the amino acids from different proteins.