Change how attention to each am ac is computed
To save computational resources, just use one list instead of two, then cast it to a tensor.
The division by the sum of the values in the attention head will be moved forward in the pipeline (see following commits).
Also delete old sum_attention function that was renamed three commits ago.