Change how avg quantities are computed - II
Switch from the previous approach - sum the quantities together and then divide by the number of proteins - to one more correct - quantities are computed in the end from the sum of att_to_am_ac of the different proteins.
This approach must still be implemented for avg_head_att_align and avg_layer_att_align.