Compute percentage of heads' attention to each amino acid
It is named avg_PH_att_to_am_ac. It differs from avg_PT_att_to_am_ac because in the first one the percentage of attention of each head is represented, while in the second one the percentage of the total attention is represented. The percentage in avg_PH_att_to_am_ac makes sense within the single heatmap, while the percentage in avg_PT_att_to_am_ac makes sense across all the heatmaps together (the sum of all the values in all the heatmaps is 100%).
Reflect the changes in parameters and returns also in __main__.