kmeans and louvain in on_set parser (#11)
* Implement manage_tot_ds.py

Move here many operations - previously done in __main__.py - involving the data structures to store the sum of quantities computed for the single peptide chains.

* Edit style

Replace number_of_layers with n_layers and number_of_heads with n_heads.

* Fix wrong broadcasting

* Edit func call and docstring

* Edit df column names

Columns representing different quantities had the same name. It was not an error but maybe a bit confusing.

* Edit logger name

Make it serious (and boring) :(

* Edit style

* Add testpaths

* Move read_pdb_files

* Ignore .out and .err

Outpute files generated by slurm in my workspace. Also remove redundant pdb_files folder.

* Remove automatic assignment of plot_path

Possible source of errors. Now the plot_path assignment is more clear, it can be done from the script and it's not based on plot_title.

* Add test dependencies

* Add pytest import mode

* Specify dependency versions

* Update README

* Delete setup.py

Switch to pyproject.toml

* Fix typo in dependency version

* Edit switched package versions

* Specify which package to install

Without it, pip detects the architecture as flat, because of the folders aimed to store the files generated by the execution of the programme. Therefore, those folders are detected as other packages of the programme and an error is produced.

* Add comment about MIN_LENGTH

* Add counter for skipped chains

Take count of the chains skipped because they don't reach the minimum number of valid residues in the chain - i.e., the total number of residues minus the number of ligands.
The counter is later used to compute the real number of chains analyzed.

* Use actual number of chains in compute_on_set.main

Now the number of chains used in compute_on_set does not include the skipped chains.
Furthermore, fix an error raised when SAMPLE_SIZE is not provided in config.txt because the list of proteins is specified in PROTEIN_CODES.

* Parse file_path instead of filename

Now I can test the class Config on a test configuration file located wherever I want, because I can specify the path every time that I create an instance of Config.

* Update docstring

* Move config.txt out of the source folder

* Adapt Config instantiation

Adapt the creation of instances of Config to the new path-based initialization.

* Add MIN_RESIDUES config option to the pipeline

* Replace info with warning for protein skip

* Log actual number of residues per chain

* Solve security issue

The issue is related to the previous version of transformers and was flagged by dependabot

* Add annotation for parse_args

* Test config_parser

* Revert commit

No simple way to stop mypy from complaining.

* try to push in my fork

* Add ignore_missing_imports in pyproject.toml

At the same time, remove ignore from the single import lines.

* Fix test files

I'm still learning :)

* returned to the previous version 05fd293 to scommit the experiment

* Bug fix

args.save_every does not exist for on_chain. I want to save everything when on_chain

* Suppress FutureWarning

Specify clean_up_tokenization spaces because leaving it blanck is deprecated now.

* added pathlib library

* structured the argparsing, as layouts -> analysis -> type of plot

* removed net_viz subparser to put all the operation under on_chain

* added reference to commit

* corrected th eimportation of generate_index_df function

* minor fix on codes

* computed the kmeans result's summary

* resolve minor logical change in the script

* complete teh preparation of the graph with nodes and edges attributes to compute the louvain analysis

* added some feature in resolution, also considering the possibility to uuse the web_groups, where the kmeans hyperparameter is given

* added resolution extraction from web_group dataframe column

* added note to not use the kmeans to generate the louvain resolution

* finalize the summary of graph for louvain communities extraction

* change for stability edge attributes -> instability edge attribute

* change stability -> instability attribute of edge in nx.graph

* added a default value to the parameter of funciton

* added the return of binary contact map for this function

* introduce the first results for binarize_contact_map in align with contact, and sum_up results for analyze session in argparse

* introduce partition result assestment summarized in sum_up script

* adjusted documentaiton

* commits on plot the 3d chain

* fixing some feature for kmeans

* added visualization parser both for pca and for netwok

* fix python version to 3.10

* removed run_protbert

* a minor change to the main argparser argument

* some logical adjustment

* minor change to function's parameters options definition

* added the attention map computing from both louvain and kmeans

* added map attention to louvain and kmeans

* net folder path added

* starting to test the code, create the parser and the respecting path folder

* added the test folder path to config parser

* added a test folder and test the correct config file

* finally found the missing parts

* test1

* test the main plot types

* fixing the get_cluster_label function work in testing

* alla results from sum_up.get_km_results and partition results works well

* corrected type error and logical error in compiler for louvain testing

* fix the problem on louvain attention map

* added two function to extract attributes from a graph, both from nodes and from edges

* minor change

* implementation of louvain for on_set parser

* implemented kmeans in on_set parser

* solved some indentations error and miscripting of louv->km variables

---------

Co-authored-by: Simone Chiarella <simonechiarella99@gmail.com>