Skip if AttributeError
Some proteins give problems in buiding the graph that we were not able to understand the reason why. Just skip the protein, and remove the corresponding code from the list.
This problem only happens with louvain, so the louvain section is moved upper. In this way the rest of the code is not executed if the error occurs in louvain.