Change download of multiple PDB codes
The Bio method is valid and faster, but often it doesn't work because the PDB FTP service is not always available (issues with public networks, VPNs, etc.).
Since the function is not working lately on my PC, I am using my own method looped over the protein codes for now.