Move fixtures to conftest.py
I figured out that those fixtures are going to be called by other fixtures that I am going to use in other test modules. Therefore, it makes more sense to put all of them in conftest.py, because they are going to be called in any case, if not directly by other fixtures.