Mock download in test_download_pdb.py
The download of the PDB files is now simulated, yet trying to stick as much as possible to the real download mechanism.
The request to the PDB download url is mocked by using responses, as well as the writing of the file, for which mock_open was implied.
All this was made in order not to rely on the external availability of the web service.

The assertion about the actual saving of the downloaded file was changed in favor of assertions about the call to the download URL -- how many time the same URL is called, the URL called is the expected one, the call is successfull.