Fix code generating NonInteractiveExampleWarning
The warning, produced by a wrong use of example(), was ignored before.
The problem is now solved by using @composite strategies instead of example() to get the dimension of the attention matrices.
Moreover, filter(lambda x: np.sum(x)<x.shape[-1]) on st_arrays is deleted because all that filtering was an unnecessary waste. It is replaced by max_value=1, which limits the execution time and is a reasonable value in terms of attention.